Move the document editor and viewer files to a separate folder
