+++ novelwriter/gui_doc/highlighter.py
@@ -33,7 +33,7 @@ from novelwriter import CONFIG, SHARED
 from novelwriter.common import checkInt, utf16CharMap
 from novelwriter.constants import nwStyles, nwUnicode
 from novelwriter.enum import nwComment
-from novelwriter.gui.doctextblock import TextBlockData
+from novelwriter.gui_doc.textblock import TextBlockData
 from novelwriter.text.formats import processComment
 from novelwriter.text.patterns import REGEX_PATTERNS, DialogParser
 from novelwriter.types import QtFontBold, QtTextUserProperty
