+++ novelwriter/gui_doc/textblock.py
