+++ novelwriter/gui_doc/viewer.py
