+++ novelwriter/gui_doc/viewerpanel.py
