+++ tests/gui_doc/test_highlighter.py
@@ -29,8 +29,8 @@ from novelwriter import CONFIG, SHARED
 from novelwriter.core.item import NWItem
 from novelwriter.core.spellcheck import NWSpellEnchant
 from novelwriter.enum import nwItemClass, nwItemLayout, nwItemType, nwTheme
-from novelwriter.gui.dochighlight import BLOCK_META, BLOCK_TITLE, GuiDocHighlighter
-from novelwriter.gui.doctextblock import TextBlockData
+from novelwriter.gui_doc.highlighter import BLOCK_META, BLOCK_TITLE, GuiDocHighlighter
+from novelwriter.gui_doc.textblock import TextBlockData
 from novelwriter.types import QtKeepAnchor
 
 R_HANDLE = "3456789abcdef"
