+++ tests/gui_doc/test_viewerpanel.py
